+++ backend/DevUp/DevUp.Domain/Identity/Repositories/IUserRepository.cs
@@ -1,6 +1,7 @@
 ﻿using DevUp.Domain.Identity.Entities;
+using DevUp.Domain.Seedwork;
 
-namespace DevUp.Infrastructure.Persistence.Repositories
+namespace DevUp.Domain.Identity.Repositories
 {
     public interface IUserRepository : IRepository<User, UserId>
     {
